requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
line-bot-sdk==3.11.0
google-cloud-vision==3.7.2
python-dotenv==1.0.1
pydantic==2.8.2